inventory handles hallu and blind
