+++ nle_code_wrapper/bot/inventory/__init__.py
@@ -1,0 +1,18 @@
+from nle_utils.item import ArmorType, ItemBeatitude, ItemClasses, ItemEnchantment, ItemErosion, ItemShopStatus
+
+from nle_code_wrapper.bot.inventory.inventory import Inventory
+from nle_code_wrapper.bot.inventory.inventory_manager import InventoryManager
+from nle_code_wrapper.bot.inventory.item import GLYPH_TO_OBJECT, Item
+
+__all__ = [
+    InventoryManager,
+    Inventory,
+    Item,
+    GLYPH_TO_OBJECT,
+    ArmorType,
+    ItemBeatitude,
+    ItemClasses,
+    ItemEnchantment,
+    ItemErosion,
+    ItemShopStatus,
+]
